+++ seven23/models/profile/models.py
@@ -31,7 +31,7 @@ class Profile(models.Model):
                                 max_length=20,
                                 choices=AVATAR_OPTIONS,
                                 help_text=_(u'Select between different origins.'))
-    auto_sync = models.BooleanField(_(u'Auto sync in app'), default=False)
+    auto_sync = models.BooleanField(_(u'Auto sync in app'), default=True)
     social_networks = models.TextField(_('social_networks blob'), blank=True, null=False)
     last_api_call = models.DateField(_(u'Last API call'),
                                 help_text=_(u'Last call on the API as a registered user'),