+++ CHANGELOG.md
@@ -25,7 +25,7 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 ### ✨ Feature
 -  Make **trial period** customizable (#98)
 -  Set **auto_Sync to True** by default (#99)
--  Make product price as Float (#100)
+-  Make product **price as Float** (#100)
 ### 🔒 Security
 - Migrate to **Django 4.2+** (#71)