+++ CHANGELOG.md
@@ -21,6 +21,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog-emoji/master/CHANGELOG.md
 -->
 
+## [1.5.0] - 2023-11-DD
+### ✨ Feature
+-  Make **trial period** customizable (#98)
+-  Set **auto_Sync to True** by default (#99)
+-  Make product **price as Float** (#100)
+-  Implement **subscription** model from stripe for Saas application (#101)
+### 🔒 Security
+- Migrate to **Django 4.2+** (#71)
+
 ## [1.4.0] - 2023-06-19
 ### 🛠 Improvements
 - Add boolean to **store if private key has been saved** and verified (#91)